requirements.txt
ccxt
tkinter
beeply
winsound
pandas
python-binance
mplfinance
